requirements.txt
numpy
netsquid
qiskit-terra
qiskit-ibmq-provider
